From Paul Martz, added support to textured light points into osgSim and OpenFlight loader, as part of the OpenFlight v16.0 spec.

+++ src/osgPlugins/OpenFlight/Pools.h
@@ -165,7 +165,7 @@ struct LPAppearance : public osg::Referenced
     float32 fadeOutDuration;
     float32 LODRangeRatio;
     float32 LODScale;
-    int32 texturePatternIndex;
+    int16 texturePatternIndex;
 };
